requirements.txt
streamlit
pandas
numpy
scikit-learn
mysql-connector-python
openpyxl  # For Excel support
ydata-profiling  # For data profiling
python-dotenv  # For environment variablesv
streamlit_pandas_profiling
pymysql